:boom: window.scrapboxに関わる型を`UserScript`から呼び出すようにし、一部の型名を書き換えた
その他、directory layoutを変えたりJSDocを書いたりした
