+++ blocks.ts
@@ -1,17 +1,21 @@
-import { Node, NodeWithoutIndent } from "./nodes.ts";
-import { Line } from "../base.ts";
+import type { Node, NodeWithoutIndent } from "./nodes.ts";
+import type { Line as LineBase } from "./base.ts";
 
-export type ParsedLine =
-  & Line
+export type Line =
+  & LineBase
   & {
     section: {
+      /** section number */
       number: number;
+      /** section開始行なら`true` */
       start: boolean;
+      /** section終了行なら`true` */
       end: boolean;
     };
   }
   & ({
-    title?: boolean;
+    /** タイトル行だったときのみ生える */
+    title?: true;
   } | {
     codeBlock: CodeBlock;
   } | {
@@ -21,12 +25,15 @@ export type ParsedLine =
   } | {
     cli: Cli;
   } | {
+    /** 番号付きリストのときのみ生える */
+    numberList?: {
+      /** 番号の長さ */
+      digit: number;
+    };
+    /** 数式を含む行のときのみ生える */
     formulaLine?: true;
-    nodes: NodeWithoutIndent[];
-  } | {
-    numberList?: { digit: number };
-    formulaLine?: true;
-    nodes: Node[];
+    /** 中に含まれるnodes */
+    nodes: Node | NodeWithoutIndent[];
   });
 
 /**  the type which represents a line in a block */
@@ -47,11 +54,17 @@ export interface TableBlock extends Block {
   /** the title of the table block */ title: string;
   /** cells included in the present line */ cells: string[];
 }
-export type Helpfeel = {
+
+/** Helpfeel記法 */
+export interface Helpfeel {
   prefix: "?";
+  /** Helpfeel本文 */
   entry: string;
-};
-export type Cli = {
+}
+
+/** Command Line記法 */
+export interface Cli {
   prefix: "$" | "%";
+  /** Command Line本文 */
   command: string;
-};
+}
