+++ mod.ts
@@ -1,5 +1,4 @@
 export * from "./api/response.ts";
 export * from "./api/error.ts";
-export * from "./userscript.ts";
-export * from "./userscript/blocks.ts";
-export * from "./userscript/nodes.ts";
+import * as UserScript from "./userscript.ts";
+export { UserScript };
